requirements.txt
asgiref==3.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
clipboard==0.0.4
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
environs==14.2.0
idna==3.10
inflection==0.5.1
marshmallow==4.0.0
packaging==25.0
PyJWT==2.9.0
pyperclip==1.9.0
python-dotenv==1.1.1
pytz==2025.2
pyvenv==0.2.2
PyYAML==6.0.2
requests==2.32.4
sqlparse==0.5.3
termcolor==1.1.0
uritemplate==4.2.0
urllib3==2.5.0
